# api/openapi.yaml
openapi: 3.0.3
info:
  title: applemusicgo API
  description: |
    HTTP API to control Apple Music on macOS via osascript (JXA / AppleScript).
    Requires Music.app on the same machine. Default port 8181 (`PORT` env).
  version: 1.0.0
  contact:
    name: applemusicgo
servers:
  - url: /
    description: Same host and port as this page (use /swagger/ on the running server)
...
paths:
...
  /now_playing:
    get:
      tags: [Player]
      summary: Current player state
      operationId: nowPlaying
      responses:
        "200":
          description: Player state JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /play:
    put:
      tags: [Player]
      summary: Play
      operationId: play
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /pause:
    put:
      tags: [Player]
      summary: Pause
      operationId: pause
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /playpause:
    put:
      tags: [Player]
      summary: Toggle play / pause
      operationId: playPause
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /stop:
    put:
      tags: [Player]
      summary: Stop
      operationId: stop
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /next:
    put:
      tags: [Player]
      summary: Next track
      operationId: next
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /previous:
    put:
      tags: [Player]
      summary: Previous track
      operationId: previous
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /seek:
    put:
      tags: [Player]
      summary: Seek to position (seconds)
      operationId: seek
      requestBody:
        $ref: "#/components/requestBodies/FormPosition"
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /volume:
    put:
      tags: [Player]
      summary: Set volume
      operationId: setVolume
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [level]
              properties:
                level:
                  type: integer
                  minimum: 0
                  maximum: 100
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /mute:
    put:
      tags: [Player]
      summary: Set mute
      operationId: setMute
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [muted]
              properties:
                muted:
                  type: string
                  description: "true or false"
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /shuffle:
    put:
      tags: [Player]
      summary: Set shuffle mode
      operationId: setShuffle
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  description: songs, albums, off, or false
                  example: songs
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /repeat:
    put:
      tags: [Player]
      summary: Set repeat mode
      operationId: setRepeat
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  description: all, one, off, or false
                  example: all
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "500":
          $ref: "#/components/responses/InternalError"
  /artwork:
    get:
      tags: [Artwork]
      summary: Current track artwork (JPEG)
      description: Fetches artwork from Music.app for the current track. Same image as `/artwork/now.jpg`.
      operationId: artworkNow
      responses:
        "200":
          description: JPEG image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /artwork/{filename}:
    get:
      tags: [Artwork]
      summary: Now-playing artwork by filename (.jpg alias)
      description: |
        Any single-segment path ending in `.jpg`, `.jpeg`, or `.png` returns the
        **current track** artwork — e.g. `/artwork/now.jpg`, `/artwork/cover.jpg`.
        Useful for media players that require a fixed `.jpg` URL. Does not apply
        to `/artwork/{artist}/{album}` (two path segments).
      operationId: artworkNowFilename
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
            example: now.jpg
      responses:
        "200":
          description: JPEG image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
...
  /notify:
    post:
      tags: [Events]
      summary: Push notification update (SSE broadcast)
      description: Used by external tools (e.g. notify.py) to feed player updates into SSE clients.
      operationId: notify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Notification"
      responses:
        "200":
          description: Accepted
        "400":
          $ref: "#/components/responses/BadRequest"
  /playlists:
    get:
      tags: [Playlists]
      summary: List user playlists
      operationId: listPlaylists
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: object
                properties:
                  playlists:
                    type: array
                    items:
                      $ref: "#/components/schemas/Playlist"
        "500":
          $ref: "#/components/responses/InternalError"
  /playlists/{id}/play:
    put:
      tags: [Playlists]
      summary: Play a playlist by id or slugified name
      operationId: playPlaylist
      parameters:
        - $ref: "#/components/parameters/PlaylistId"
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /library/artists:
    get:
      tags: [Library]
      summary: Browse artists
      operationId: libraryArtists
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/LimitArtists"
      responses:
        "200":
          description: Paginated artists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistsPage"
        "500":
          $ref: "#/components/responses/InternalError"
  /library/artists/{artist}/albums:
    get:
      tags: [Library]
      summary: Albums by artist
      operationId: libraryArtistAlbums
      parameters:
        - $ref: "#/components/parameters/Artist"
      responses:
        "200":
          description: Artist albums
          content:
            application/json:
              schema:
                type: object
                properties:
                  artist:
                    type: string
                  albums:
                    type: array
                    items:
                      $ref: "#/components/schemas/AlbumRef"
        "500":
          $ref: "#/components/responses/InternalError"
  /library/artists/{artist}/tracks:
    get:
      tags: [Library]
      summary: Tracks by artist
      operationId: libraryArtistTracks
      parameters:
        - $ref: "#/components/parameters/Artist"
      responses:
        "200":
          description: Artist tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistTracks"
        "500":
          $ref: "#/components/responses/InternalError"
  /library/artists/{artist}/play:
    put:
      tags: [Library]
      summary: Play all tracks by artist
      operationId: playArtist
      parameters:
        - $ref: "#/components/parameters/Artist"
      responses:
        "200":
          description: Updated player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
...
    PlaylistId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  requestBodies:
    FormPosition:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            type: object
            required: [position]
            properties:
              position:
                type: number
                format: float
                description: Position in seconds
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
    InternalError:
      description: Internal server error
  schemas:
...
    PlayerState:
      type: object
      description: Current player state (fields vary when stopped)
      properties:
        player_state:
          type: string
          enum: [playing, paused, stopped]
        id:
          type: string
        name:
          type: string
        artist:
          type: string
        album:
          type: string
        volume:
          type: integer
        muted:
          type: boolean
        repeat:
          type: string
        shuffle:
          type: boolean
        player_position:
          type: number
        player_duration:
          type: number
        position_timestamp:
          type: integer
          format: int64
        playlist:
          type: string
...